+++ electron/src/electron-utils/utilsDrop.ts
@@ -50,6 +50,8 @@ export class UtilsDrop {
    */
   public async dropElements(db: any, type: string): Promise<void> {
     let msg = '';
+    let stmt1 = `AND name NOT LIKE ('sqlite_%')`;
+
     switch (type) {
       case 'index':
         msg = 'DropIndexes';
@@ -59,6 +61,7 @@ export class UtilsDrop {
         break;
       case 'table':
         msg = 'DropTables';
+        stmt1 += ` AND name NOT IN ('sync_table')`;
         break;
       case 'view':
         msg = 'DropViews';
@@ -68,7 +71,7 @@ export class UtilsDrop {
     }
     // get the element's names
     let stmt = 'SELECT name FROM sqlite_master WHERE ';
-    stmt += `type = '${type}' AND name NOT LIKE 'sqlite_%';`;
+    stmt += `type = '${type}' ${stmt1};`;
     try {
       const elements: any[] = await this._uSQLite.queryAll(db, stmt, []);
       if (elements.length > 0) {