+++ ios/Plugin/CapacitorSQLitePlugin.m
@@ -33,6 +33,7 @@ CAP_PLUGIN(CapacitorSQLitePlugin, "CapacitorSQLite",
            CAP_PLUGIN_METHOD(isNCDatabase, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(isTableExists, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(getDatabaseList, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(getTableList, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(getMigratableDbList, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(addSQLiteSuffix, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(deleteOldDatabases, CAPPluginReturnPromise);